+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [v3.0.3](https://github.com/malcommac/SwiftDate/releases/tag/3.0.3) (2016-)
+[Full Changelog](https://github.com/malcommac/SwiftDate/compare/3.0.2..v3.0.3)
+- Deprecated .toISO8101String(). Use .toString(.ISO8601) instead ([Commit](https://github.com/malcommac/SwiftDate/commit/9451ed3fcf3ba996a347ea96ce45694371aa4f3a))
+- Locale is not set to calendar object but inside the formatter ([Commit](https://github.com/malcommac/SwiftDate/commit/0ccc172d506034f1cd1b949100fbfca3f16bf303))
+
 ## [v3.0.1](https://github.com/malcommac/SwiftDate/releases/tag/3.0.1) (2016-01-28)
 [Full Changelog](https://github.com/malcommac/SwiftDate/compare/3.0.0..v3.0.1)
 - Fixed Carthage build (SwiftDate now compiles on iOS,TvOS,OSX and WatchOS)